--- .project
@@ -1,13 +1,0 @@
-<projectDescription>
-  <name>main</name>
-  <buildSpec>
-    <buildCommand>
-      <name>org.scala-ide.sdt.core.scalabuilder</name>
-    </buildCommand>
-  </buildSpec>
-  <natures>
-    <nature>org.scala-ide.sdt.core.scalanature</nature>
-    <nature>org.eclipse.jdt.core.javanature</nature>
-  </natures>
-  <linkedResources> </linkedResources>
-</projectDescription>
