+++ core/src/main/scala/org/influxdb/scala/HTTPServiceComponent.scala
