+++ core/src/main/scala/org/influxdb/scala/InfluxDBAdminAPI.scala
@@ -1,0 +1,10 @@
+package org.influxdb.scala
+
+import scala.concurrent.Future
+import org.influxdb.scala.HTTPServiceComponent.HTTPService
+
+trait InfluxDBAdminAPI {
+
+  self: HTTPService =>
+  def listDatabases: Future[List[String]]
+}
