+++ core/src/main/scala/org/influxdb/scala/Precision.scala
