+++ core/src/main/scala/org/influxdb/scala/Series.scala
