+++ core/src/main/scala/org/influxdb/scala/TSeries.scala
