+++ core/src/main/scala/org/influxdb/scala/package.scala
