+++ standalone/src/main/scala/org/influxdb/scala/AsyncHttpClientImpl.scala
